+++ src/main/java/org/platanus/webboard/controller/user/RoleService.java
@@ -1,31 +1,16 @@
 package org.platanus.webboard.controller.user;
 
-import lombok.RequiredArgsConstructor;
 import org.platanus.webboard.domain.Role;
-import org.platanus.webboard.domain.RoleRepository;
 import org.platanus.webboard.domain.User;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class RoleService {
-    private final RoleRepository roleRepository;
+public interface RoleService {
+    Role join(Role role);
 
-    public Role join(Role role) {
-        return roleRepository.save(role);
-    }
+    int remove(Role role);
 
-    public int remove(Role role) {
-        return roleRepository.delete(role.getRole(), role.getUserId());
-    }
+    List<Role> findAll();
 
-    public List<Role> findAll() {
-        return roleRepository.findAll();
-    }
-
-    public List<Role> findByUser(User user) {
-        return roleRepository.findByUserId(user.getId());
-    }
+    List<Role> findByUser(User user);
 }
